requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
requests>=2.31.0
pytest>=7.0.0
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pytest-cov>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
scipy>=1.10.0
httpx>=0.24.0
pytest-asyncio>=0.21.0 